backend/requirements.txt
Flask>=3.0
Flask-SocketIO>=5.3
flask-cors
psycopg2-binary
PyJWT>=2.8
python-dotenv
gunicorn
email-validator
bcrypt
